refactor: 💡 fetching photos
